mbbsd: Drop user_has_email()
user_has_email() calls normalize_email() on a copy of the user's email
address, which then does str_lower() on that copy. This is unnecessary,
since we only want to check the validaty of the email address.

Replace all invocations with a direct call to is_valid_email() and drop
user_has_email().